requirements.txt
web3==6.17.2
python-dotenv==1.0.1